ajout google trad et spotify
